requirements.txt
Django==5.2.1
Pillow==11.2.1
djangorestframework==3.14.0
